requirements.txt
Django==3.1.5
mysqlclient==2.0.3
yamjam==0.1.7
djangorestframework==3.12.2
requests~=2.22.0